Playing around some more again

+++ src/main/kotlin/com/cinchfinancial/kool/profile/MXAttributes.kt
@@ -1,5 +1,7 @@
 package com.cinchfinancial.kool.profile
 
+import com.cinchfinancial.kool.types.Numeric
+
 /**
  * Created by mark on 5/2/17.
  */
@@ -15,8 +17,8 @@ class MXAttributes(properties: Map<String, Any?>) : BaseAttributes(properties) {
         val last_refresh_on: String by scalarValue(properties)
         val newest_transaction_on: String by scalarValue(properties)
         val oldest_transaction_on: String by scalarValue(properties)
-        val months: Int by scalarValue(properties)
-        val days: Int by scalarValue(properties)
+        val months: Numeric by scalarValue(properties)
+        val days: Numeric by scalarValue(properties)
 
     }
 }
